backend/requirements.txt
fastapi
uvicorn
python-multipart
python-dotenv
openai
ollama
requests